[refactor] apps minor restructure for redux and react-redux installs
