+++ image_editor/static/js/api/facebookApi.js
