+++ image_editor/static/js/api/imageApi.js
@@ -104,6 +104,7 @@ const ImageApi = {
   */
   uploadImage(files) {
     files.forEach(file => {
+      console.log(file, 'the single file')
       this.setState({filename: file.name});
       let reader = new FileReader();
       reader.readAsDataURL(file);
