+++ package.json
@@ -20,8 +20,10 @@
     "react": "^0.14.3",
     "react-dom": "^0.14.3",
     "react-dropzone": "^3.3.0",
+    "react-redux": "^4.4.5",
     "react-responsive-carousel": "^2.0.2",
     "react-slick": "^0.9.3",
+    "redux": "^3.5.2",
     "sass-loader": "^3.1.2",
     "superagent": "^1.6.1",
     "superagent-django-csrf": "^0.1.3",
